requirements.txt
py-cord
google-genai
aiohttp
psutil
requests
audioop-lts
yt_dlp
ytmusicapi
PyNaCl
tomli-w
selenium
webdriver-manager
beautifulsoup4
